Adapt to Master
