--- app/Resources/pdf/wkhtmltopdf/include/wkhtmltox/dllend.inc
@@ -1,28 +1,0 @@
-/*
- * Copyright 2010 wkhtmltopdf authors
- *
- * This file is part of wkhtmltopdf.
- *
- * wkhtmltopdf is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * wkhtmltopdf is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with wkhtmltopdf.  If not, see <http: *www.gnu.org/licenses/>.
- */
-
-#ifdef __WKHTMLTOPDF_DLLBEGIN__
-
-#undef __WKHTMLTOPDF_DLLBEGIN__
-#undef DLL_PUBLIC
-#undef DLL_LOCAL
-#undef CAPI
-#undef CALLTYPE
-
-#endif /*__WKHTMLTOPDF_DLLBEGIN__*/
